requirements.txt
boto3==1.26.45
botocore==1.29.45
eventlet==0.33.2
Flask==2.2.2
mysql-connector-python==8.0.31
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
s3transfer==0.6.0
Werkzeug==2.2.2

